feat(backend): filtering for GET all quizzes, PUT and DELETE for quizzes

+++ backend/src/main/java/com/quizzard/quizzard/service/QuizService.java
@@ -1,11 +1,14 @@
 package com.quizzard.quizzard.service;
 
+import com.quizzard.quizzard.exception.AccessDeniedException;
+import com.quizzard.quizzard.exception.ResourceNotFoundException;
 import com.quizzard.quizzard.model.Question;
 import com.quizzard.quizzard.model.Quiz;
 import com.quizzard.quizzard.model.User;
 import com.quizzard.quizzard.model.request.CreateQuizRequest;
 import com.quizzard.quizzard.model.request.QuestionRequest;
 import com.quizzard.quizzard.model.request.SolveQuizRequest;
+import com.quizzard.quizzard.model.request.UpdateQuizRequest;
 import com.quizzard.quizzard.model.response.QuestionResponse;
 import com.quizzard.quizzard.model.response.QuizResponse;
 import com.quizzard.quizzard.model.response.SolveQuizResponse;
@@ -75,9 +78,28 @@ public class QuizService {
         return mapQuizToQuizResponse(quiz);
     }
 
-    // List all quizzes
-    public List<QuizResponse> getAllQuizzes() {
-        return mapQuizzesToQuizResponses(quizRepository.findAll());
+    public Object getAllQuizzes(Optional<String> username, Optional<Integer> minDifficulty, Optional<Integer> maxDifficulty) {
+        if (username.isPresent() && minDifficulty.isPresent() && maxDifficulty.isPresent()) {
+            User user = userService.getOneUserByUsername(username.get());
+            return mapQuizzesToQuizResponses(quizRepository.findByAuthorAndDifficultyBetween(user, minDifficulty.get().doubleValue(), maxDifficulty.get().doubleValue()));
+        } else if (username.isPresent() && minDifficulty.isPresent()) {
+            User user = userService.getOneUserByUsername(username.get());
+            return mapQuizzesToQuizResponses(quizRepository.findByAuthorAndDifficultyGreaterThanEqual(user, minDifficulty.get().doubleValue()));
+        } else if (username.isPresent() && maxDifficulty.isPresent()) {
+            User user = userService.getOneUserByUsername(username.get());
+            return mapQuizzesToQuizResponses(quizRepository.findByAuthorAndDifficultyLessThanEqual(user, maxDifficulty.get().doubleValue()));
+        } else if (minDifficulty.isPresent() && maxDifficulty.isPresent()) {
+            return mapQuizzesToQuizResponses(quizRepository.findByDifficultyBetween(minDifficulty.get().doubleValue(), maxDifficulty.get().doubleValue()));
+        } else if (username.isPresent()) {
+            User user = userService.getOneUserByUsername(username.get());
+            return mapQuizzesToQuizResponses(quizRepository.findByAuthor(user));
+        } else if (minDifficulty.isPresent()) {
+            return mapQuizzesToQuizResponses(quizRepository.findByDifficultyGreaterThanEqual(minDifficulty.get().doubleValue()));
+        } else if (maxDifficulty.isPresent()) {
+            return mapQuizzesToQuizResponses(quizRepository.findByDifficultyLessThanEqual(maxDifficulty.get().doubleValue()));
+        } else {
+            return mapQuizzesToQuizResponses(quizRepository.findAll());
+        }
     }
 
     // Find specific quiz with its ID
@@ -90,26 +112,33 @@ public class QuizService {
     }
 
     // Update quiz
-    public Quiz updateQuiz(Long id, Quiz updatedQuiz) {
+    public QuizResponse updateQuiz(String username, Long id, UpdateQuizRequest updatedQuiz) {
         Optional<Quiz> quizOptional = quizRepository.findById(id);
-        if (quizOptional.isPresent()) {
-            Quiz existingQuiz = quizOptional.get();
-            existingQuiz.setTitle(updatedQuiz.getTitle());
-            existingQuiz.setDescription(updatedQuiz.getDescription());
-            existingQuiz.setImage(updatedQuiz.getImage());
-            existingQuiz.setDifficulty(updatedQuiz.getDifficulty());
-            existingQuiz.setUpdatedAt(new java.util.Date());
-            return quizRepository.save(existingQuiz);
-        } else {
-            return null;  // Or handle the case where the quiz doesn't exist
-        }
+        if (quizOptional.isEmpty())
+            throw new ResourceNotFoundException("Quiz not found with id " + id);
+        Quiz quiz = quizOptional.get();
+        User user = userService.getOneUserByUsername(username);
+        if (quiz.getAuthor().getId() != user.getId())
+            throw new AccessDeniedException("You are not the author of this quiz");
+        if(updatedQuiz.getTitle() != null)
+            quiz.setTitle(updatedQuiz.getTitle());
+        if(updatedQuiz.getDescription() != null)
+            quiz.setDescription(updatedQuiz.getDescription());
+        if(updatedQuiz.getImage() != null)
+            quiz.setImage(updatedQuiz.getImage());
+        quizRepository.save(quiz);
+        return mapQuizToQuizResponse(quiz);
     }
 
     // Delete a quiz
-    public void deleteQuiz(Long id) {
+    public void deleteQuiz(String username ,Long id) {
+        User user = userService.getOneUserByUsername(username);
+        Optional<Quiz> quiz = quizRepository.findById(id);
+        if (quiz.isEmpty())
+            throw new ResourceNotFoundException("Quiz not found with id " + id);
+        if (quiz.get().getAuthor().getId() != user.getId())
+            throw new AccessDeniedException("You are not the author of this quiz");
         quizRepository.deleteById(id);
     }
-
-
 
 }
